+++ CHANGES.md
@@ -8,6 +8,7 @@ NEXT VERSION
 **Deprecated:**
 
 - Deprecated `containsString`; use `containsSubstring` instead. `containsString` clashes with an NSString method introduced in iOS 8.
+- Deprecated `assertThatAfter`/`futureValueOf`. Use `assertWithTimeout`/`thatEventually` instead.
 
 
 Version 4.1.1